GameLevel Class
I have added the GameLevel class. I believe this class can be helpful for the logic of the game. The basic idea, GameLevel constructor generates a random number array listing depending on seed value, level, cycle iterations.  The levels are integer values with would allow the game to have 1 to n buttons. The cycles would allow the game to increase the sequence of inputs.